+++ lib/models/player.dart
@@ -8,17 +8,17 @@ class Player {
   String league;
   int number;
   String injured;
-  Stats stats;
+  PlayerStats stats;
 
   Player(this.id, this.name, this.age, this.position, this.team, this.teamName, this.league, this.number, this.injured, this.stats);
 }
 
-class Stats {
+class PlayerStats {
   int appearences;
   int goals;
   int assits;
   int yellowCards;
   int redCards;
 
-  Stats(this.appearences, this.goals, this.assits, this.yellowCards, this.redCards);
+  PlayerStats(this.appearences, this.goals, this.assits, this.yellowCards, this.redCards);
 }