+++ lib/services/data_service.dart
@@ -62,5 +62,4 @@ class DataService {
       return [];
     }
   }
-
 }