+++ lib/util/themes.dart
@@ -1,9 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:gfg_mobile/services/auth_service.dart';
 
 class Themes {
   Themes();
-  final auth = new AuthService();
   final backgroundColor = Colors.grey[850];
   final primaryColor = Colors.black;
   final buttonColor = Colors.black;
@@ -26,9 +24,14 @@ class Themes {
     ),
   );
 
+  Icon profileIcon = new Icon(Icons.account_circle, color: Colors.limeAccent[700],);
+  Icon playersIcon = new Icon(Icons.face, color: Colors.limeAccent[700],);
+  Icon charityIcon = new Icon(Icons.favorite, color: Colors.limeAccent[700],);
+  Icon subsIcon = new Icon(Icons.format_list_bulleted, color: Colors.limeAccent[700],);
+  Icon aboutIcon = new Icon(Icons.text_format, color: Colors.limeAccent[700],);
   Icon searchIcon = new Icon(Icons.search, color: Colors.limeAccent[700],);
-
   Icon backIcon = new Icon(Icons.arrow_back, color: Colors.limeAccent[700],);
+  Icon matchesIcon = new Icon(Icons.calendar_today, color: Colors.limeAccent[700]);
 
   Text logout = new Text('Logout', style: new TextStyle(fontSize: 17.0, color: Colors.limeAccent[700]));
 }